+++ spec/test/configuration_spec.rb
@@ -11,7 +11,7 @@ RSpec.describe OpenapiFirst::Test::Configuration do
 
   describe 'ignored_unknown_status' do
     it 'has a default value' do
-      expect(configuration.ignored_unknown_status).to eq [404]
+      expect(configuration.ignored_unknown_status).to contain_exactly(401, 404, 500)
     end
 
     it 'can be extended' do